Remove unnecessary float cast

+++ scripts/player/TongueSpring.cs
@@ -11,7 +11,7 @@ public partial class TongueSpring : DampedSpringJoint2D
 		NodeB = Target.GetPath(); // Set secondary spring node to target
 
 		LookAt(Target.GlobalPosition);
-		Rotate((float)-0.5f * Mathf.Pi); // NECESSARY, but no clue why :3
+		Rotate(-0.5f * Mathf.Pi); // NECESSARY, but no clue why :3
 		Length = LengthFactor * GlobalPosition.DistanceTo(Target.GlobalPosition);
 	}
 }
